requirements.txt
segno>=1.5.0
Pillow>=9.0.0
